modyle5-task1: * task class
