--- src/view/task/common.ts
@@ -1,6 +1,0 @@
-enum TaskTemplateMode {
-  PREVIEW = `preview`,
-  EDIT = `edit`,
-}
-
-export { TaskTemplateMode };
